Refactor orchestration logic to kvcache backends
